rename components to index.tsx
