+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -25,22 +25,22 @@ import { ConnectViaEthernet } from '@opentrons/app/src/pages/ConnectViaEthernet'
 import { ConnectViaUSB } from '@opentrons/app/src/pages/ConnectViaUSB/ConnectViaUSB'
 import { ConnectViaWifi } from '@opentrons/app/src/pages/ConnectViaWifi'
 import { EmergencyStop } from '../pages/EmergencyStop'
-import { NameRobot } from '../pages/OnDeviceDisplay/NameRobot'
-import { NetworkSetupMenu } from '../pages/NetworkSetupMenu/NetworkSetupMenu'
+import { NameRobot } from '../pages/NameRobot'
+import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
 import { ProtocolSetup } from '@opentrons/app/src/pages/ProtocolSetup'
 import { TempODDMenu } from '../pages/OnDeviceDisplay/TempODDMenu'
 import { RobotDashboard } from '../pages/RobotDashboard'
 import { RobotSettingsDashboard } from '@opentrons/app/src/pages/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../pages/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/ProtocolDetails'
-import { RunningProtocol } from '../pages/RunningProtocol/RunningProtocol'
-import { RunSummary } from '../pages/RunSummary/RunSummary'
+import { RunningProtocol } from '../pages/RunningProtocol'
+import { RunSummary } from '../pages/RunSummary'
 import { UpdateRobot } from '../pages/UpdateRobot/UpdateRobot'
 import { UpdateRobotDuringOnboarding } from '../pages/UpdateRobot/UpdateRobotDuringOnboarding'
-import { InstrumentsDashboard } from '../pages/InstrumentsDashboard/InstrumentsDashboard'
+import { InstrumentsDashboard } from '../pages/InstrumentsDashboard'
 import { InstrumentDetail } from '@opentrons/app/src/pages/InstrumentDetail'
-import { Welcome } from '../pages/OnDeviceDisplay/Welcome'
-import { InitialLoadingScreen } from '../pages/InitialLoading/InitialLoadingScreen'
+import { Welcome } from '../pages/Welcome'
+import { InitialLoadingScreen } from '../pages/InitialLoadingScreen'
 import { DeckConfigurationEditor } from '../pages/DeckConfiguration'
 import { PortalRoot as ModalPortalRoot } from './portal'
 import { getOnDeviceDisplaySettings, updateConfigValue } from '../redux/config'
