+++ app/src/resources/networking/__tests__/useNetworkConnection.test.tsx
